requirements.txt
asgiref==3.4.1
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
cssselect2==0.4.1
Django==3.2.7
django-crispy-forms==1.13.0
fonttools==4.27.1
html5lib==1.1
idna==3.2
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
pydyf==0.1.1
PyJWT==2.1.0
pyphen==0.11.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.0
twilio==7.0.0
urllib3==1.26.7
weasyprint==53.3
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.1.9
